requirements.txt
# Django and REST framework
Django==4.2.0
djangorestframework==3.14.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==23.2

# Celery and task management
django-celery-beat==2.5.0
django-celery-results==2.5.0
celery==5.3.0
redis==4.5.5

# Web scraping and networking
requests==2.30.0
beautifulsoup4==4.12.0
python-nmap==0.7.1
pyOpenSSL==23.1.1
dnspython==2.3.0
python-whois==0.8.0

# PDF generation
weasyprint==59.0
reportlab==3.6.12

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.4.0

# Development tools
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.5
black==23.3.0
flake8==6.0.0

